Revert "Guest players"
