+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "format": "prettier \"src/**/*.{js,html}\" --write",
     "lint": "eslint \"src/**/*.{js,jsx}\" --quiet",
-    "dev": "parcel src/index.html",
-    "build": "parcel build src/index.html",
+      "dev": "parcel src/index.html",
+      "build": "parcel build src/index.html",
     "test": "echo \"Error: no test specified\" && exit 1",
     "preinstall": "npx npm-force-resolutions"
   },
@@ -26,11 +26,8 @@
     "prettier": "^2.1.2"
   },
   "dependencies": {
-    "@reach/router": "^1.3.4",
-    "bootstrap": "^4.5.3",
     "gh-pages": "^3.1.0",
     "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
     "react-scripts": "^4.0.0"
   }
