+++ src/Selector.js
@@ -1,58 +1,79 @@
-import React, { useState } from "react";
+import React, { useReducer} from "react";
+import Teams from "./TeamData.json";
 import Players from "./Players.json";
+import FixtureDetails from "./FixtureDetails";
+import PlayerList from "./PlayerList";
+import Result from "./Result";
+import Officials from "./Officials";
+import reducer from "./Reducer";
 
-const Rounds = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18];
+  const initialState = {
+    round: 1,
+    team: "Adelaide Spurs",
+    selected: [],
+  };
 
-const Selector = ({round, team, guests, selected, dispatch}) => {
+  const Rounds = [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18];
 
-  var d = new Date();
-  var n = d.getTime();
-  
-  console.log("Running Selector(",n,"). guests:", guests.length);
+const Selector = () => {
 
-  var availablePlayers = Players;
-  availablePlayers = availablePlayers.concat(guests);
-  availablePlayers = availablePlayers.filter(extractPlayersByTeam).sort((a, b) => a.name > b.name);
-  
+  const [state, dispatch] = useReducer(reducer, initialState);
+  const {round, team, selected} = state;
+
+  var availablePlayers = Players.filter(extractPlayersByTeam).sort((a, b) => a.name > b.name);
+  var selectedPlayers = availablePlayers;
+
+
   function extractPlayersByTeam(plr) {
     return plr.team === team;    
   }
 
-  function updateSelected (selection) {
-
-    var updatedSelections = [];
-    updatedSelections = updatedSelections.concat(selected);
-
-    console.log("selection: ", selection);
-    console.log("Selected Before:",selected);
-
-    if (!updatedSelections.includes(selection)) {
-      updatedSelections[updatedSelections.length] = selection;
-    } else {
-      var index = updatedSelections.indexOf(selection);
-      updatedSelections.splice(index, 1);
-    }
-
-
-    dispatch({type: 'CHANGE_SELECTED', payload: updatedSelections});
-    console.log("Selected After:",selected);
-  }
-
   return (
     <div className="teamsheet-generator">
-    {/* <h3 >{team} (Round {round})</h3> */}
-      <h3 >{team}</h3>
       <table>
         <tbody>
           <tr valign="top">
             <td> 
-              <div >
+              <div className="search-params">
                 <form>
+                  <label htmlFor="round">
+                    Round
+                    <select
+                    id="round"
+                    value={round}
+                      placeholder="Round"
+                      onChange={(e) => dispatch({type: 'CHANGE_ROUND', payload: e.target.value})}
+                      onBlur={(e) => dispatch({type: 'CHANGE_ROUND', payload: e.target.value})}
+                      >
+                        {
+                        Rounds.map(round => (
+                          <option key={round} value={round}>{round}</option>
+                        ))
+                      }
+
+                      </select>
+                  </label>
+
+                  <label htmlFor="team">
+                    Team
+                    <select 
+                      id="team"
+                      value={team}
+                      onChange={(e) => dispatch({type: 'CHANGE_TEAM', payload: e.target.value})}
+                      onBlur={(e) => dispatch({type: 'CHANGE_TEAM', payload: e.target.value})}
+                    >
+                      {
+                      Teams.map(team => (
+                        <option key={team.id} value={team.name}>{team.name}</option>))
+                      }
+                    </select>
+                  </label>
+
                   <table>
                     <thead>
                       <tr>
                         <th>Name</th>
-                        <th></th>
+                        <th>playing</th>
                       </tr>
                     </thead>
                     <tbody>
@@ -66,21 +87,36 @@ const Selector = ({round, team, guests, selected, dispatch}) => {
                             <input 
                             type="checkbox" checked={p.active}
                             id={p.id} 
-                            onClick={() => updateSelected(p)}
+                            onClick={(e) => dispatch({type: 'TOGGLE_PLAYER', payload: p, extra: e.target.checked})}
                             />
               
                           </td>
                         </tr>
-                      ))
+                    ))
                     }
 
                     </tbody>
                   </table>
+
+                  {/* <br/>
+                  <button>Print</button> */}
                 </form>
               </div> 
             </td>
             <td>
-              {/* <PrintPreview round={round} team={team} selected={selected}/> */}
+              <div className="printArea" >
+                <FixtureDetails
+                  round={round}
+                  team={team}
+                />
+                <PlayerList team={team} players={selected} />
+                <br />
+                <Result 
+                  round={round}
+                  team={team} />
+                  <br />
+                  <Officials />
+              </div>
             </td>
           </tr>
         </tbody>
